Fix Licenses issues in runtime
